Add custom authentication routes and controller
